+++ packages/rpc-provider/src/substrate-connect/index.spec.ts
@@ -23,6 +23,7 @@ import type { ScProviderClass } from '.';
 import type { HealthChecker, SmoldotHealth } from './Health';
 
 import { jest } from '@jest/globals';
+import * as ScConnect from '@substrate/connect';
 
 const wait = (ms: number) => new Promise((res) => setTimeout(res, ms));
 
@@ -208,7 +209,7 @@ beforeAll(async () => {
 describe('ScProvider', () => {
   describe('on', () => {
     it('emits `connected` as soon as the chain is not syncing', async () => {
-      const provider = new ScProvider('');
+      const provider = new ScProvider(ScConnect, '');
 
       await provider.connect(undefined, mockedHealthChecker.healthChecker);
 
@@ -222,7 +223,7 @@ describe('ScProvider', () => {
     });
 
     it('stops receiving notifications after unsubscribing', async () => {
-      const provider = new ScProvider('');
+      const provider = new ScProvider(ScConnect, '');
 
       await provider.connect(undefined, mockedHealthChecker.healthChecker);
 
@@ -236,7 +237,7 @@ describe('ScProvider', () => {
     });
 
     it('synchronously emits connected if the Provider is already `connected`', async () => {
-      const provider = new ScProvider('');
+      const provider = new ScProvider(ScConnect, '');
 
       await provider.connect(undefined, mockedHealthChecker.healthChecker);
 
@@ -249,7 +250,7 @@ describe('ScProvider', () => {
     });
 
     it('emits `disconnected` once the chain goes back to syncing', async () => {
-      const provider = new ScProvider('');
+      const provider = new ScProvider(ScConnect, '');
 
       await provider.connect(undefined, mockedHealthChecker.healthChecker);
 
@@ -275,7 +276,7 @@ describe('ScProvider', () => {
 
   describe('hasSubscriptions', () => {
     it('supports subscriptions', async () => {
-      const provider = new ScProvider('');
+      const provider = new ScProvider(ScConnect, '');
 
       await provider.connect(undefined, mockedHealthChecker.healthChecker);
 
@@ -285,7 +286,7 @@ describe('ScProvider', () => {
 
   describe('clone', () => {
     it('can not be clonned', async () => {
-      const provider = new ScProvider('');
+      const provider = new ScProvider(ScConnect, '');
 
       await provider.connect(undefined, mockedHealthChecker.healthChecker);
 
@@ -295,7 +296,7 @@ describe('ScProvider', () => {
 
   describe('connect', () => {
     it.skip('does not create a new chain when trying to re-connect while the current chain is syncing', async () => {
-      const provider = new ScProvider('');
+      const provider = new ScProvider(ScConnect, '');
 
       await provider.connect(undefined, mockedHealthChecker.healthChecker);
       const chain = mockedConnector.latestChain();
@@ -305,7 +306,7 @@ describe('ScProvider', () => {
     });
 
     it('throws when trying to connect on an already connected Provider', async () => {
-      const provider = new ScProvider('');
+      const provider = new ScProvider(ScConnect, '');
 
       await provider.connect(undefined, mockedHealthChecker.healthChecker);
 
@@ -319,7 +320,7 @@ describe('ScProvider', () => {
 
   describe('disconnect', () => {
     it.skip('removes the chain and cleans up', async () => {
-      const provider = new ScProvider('');
+      const provider = new ScProvider(ScConnect, '');
 
       await provider.connect(undefined, mockedHealthChecker.healthChecker);
       const chain = mockedConnector.latestChain();
@@ -330,7 +331,7 @@ describe('ScProvider', () => {
     });
 
     it('does not throw when disconnecting on an already disconnected Provider', async () => {
-      const provider = new ScProvider('');
+      const provider = new ScProvider(ScConnect, '');
 
       await provider.connect(undefined, mockedHealthChecker.healthChecker);
 
@@ -341,7 +342,7 @@ describe('ScProvider', () => {
 
   describe('send', () => {
     it('throws when trying to send a request while the Provider is not connected', async () => {
-      const provider = new ScProvider('');
+      const provider = new ScProvider(ScConnect, '');
 
       await provider.connect(undefined, mockedHealthChecker.healthChecker);
 
@@ -349,7 +350,7 @@ describe('ScProvider', () => {
     });
 
     it.skip('receives responses to its requests', async () => {
-      const provider = new ScProvider('');
+      const provider = new ScProvider(ScConnect, '');
 
       await provider.connect(undefined, mockedHealthChecker.healthChecker);
       const chain = mockedConnector.latestChain();
@@ -377,7 +378,7 @@ describe('ScProvider', () => {
     });
 
     it.skip("rejects when the response can't be deserialized", async () => {
-      const provider = new ScProvider('');
+      const provider = new ScProvider(ScConnect, '');
 
       await provider.connect(undefined, mockedHealthChecker.healthChecker);
       const chain = mockedConnector.latestChain();
@@ -395,7 +396,7 @@ describe('ScProvider', () => {
     });
 
     it.skip('rejects when the smoldot chain has crashed', async () => {
-      const provider = new ScProvider('');
+      const provider = new ScProvider(ScConnect, '');
 
       await provider.connect(undefined, mockedHealthChecker.healthChecker);
       const chain = mockedConnector.latestChain();
@@ -416,7 +417,7 @@ describe('ScProvider', () => {
 
   describe('subscribe', () => {
     it.skip('subscribes and recives messages until it unsubscribes', async () => {
-      const provider = new ScProvider('');
+      const provider = new ScProvider(ScConnect, '');
 
       await provider.connect(undefined, mockedHealthChecker.healthChecker);
       const chain = mockedConnector.latestChain();
@@ -481,7 +482,7 @@ describe('ScProvider', () => {
     });
 
     it.skip('ignores subscription messages that were received before the subscription token', async () => {
-      const provider = new ScProvider('');
+      const provider = new ScProvider(ScConnect, '');
 
       await provider.connect(undefined, mockedHealthChecker.healthChecker);
       const chain = mockedConnector.latestChain();
@@ -519,7 +520,7 @@ describe('ScProvider', () => {
     });
 
     it.skip('emits the error when the message has an error', async () => {
-      const provider = new ScProvider('');
+      const provider = new ScProvider(ScConnect, '');
 
       await provider.connect(undefined, mockedHealthChecker.healthChecker);
       const chain = mockedConnector.latestChain();
@@ -561,7 +562,7 @@ describe('ScProvider', () => {
     });
 
     it('errors when subscribing to an unsupported method', async () => {
-      const provider = new ScProvider('');
+      const provider = new ScProvider(ScConnect, '');
 
       await provider.connect(undefined, mockedHealthChecker.healthChecker);
 
@@ -576,7 +577,7 @@ describe('ScProvider', () => {
 
   describe('unsubscribe', () => {
     it('rejects when trying to unsubscribe from un unexisting subscription', async () => {
-      const provider = new ScProvider('');
+      const provider = new ScProvider(ScConnect, '');
 
       await provider.connect(undefined, mockedHealthChecker.healthChecker);
 
@@ -589,7 +590,7 @@ describe('ScProvider', () => {
   });
 
   it.skip('cleans up the stale subscriptions once it reconnects', async () => {
-    const provider = new ScProvider('');
+    const provider = new ScProvider(ScConnect, '');
 
     await provider.connect(undefined, mockedHealthChecker.healthChecker);
     const chain = mockedConnector.latestChain();